+++ fixtures/js-stubs/types.tsx
