+++ tests/js/sentry-test/loadFixtures.ts
@@ -4,7 +4,7 @@
 import fs from 'fs';
 import path from 'path';
 
-import TestStubFixtures from '../../../fixtures/js-stubs/types.test';
+import TestStubFixtures from '../../../fixtures/js-stubs/types';
 
 const FIXTURES_ROOT = path.join(__dirname, '../../../fixtures');
